+++ src/profile.c
@@ -48,14 +48,14 @@ profile_init(profile_t *p)
 static int
 profile_dump(profile_t *p, UNUSED void *udata)
 {
-    printf("%s: n=%ld min=%ld avg=%Lf max=%ld total=%Lf\n", p->name, p->n, p->min, p->avg, p->max, p->n * p->avg);
+    printf("%s: n=%ld min=%ld avg=%Lf max=%ld total=%Lf\n", p->name, (long)p->n, (long)p->min, p->avg, (long)p->max, p->n * p->avg);
     return 0;
 }
 
 static int
 profile_dump_sec(profile_t *p, UNUSED void *udata)
 {
-    printf("%s: n=%ld min=%Lf avg=%Lf max=%Lf total=%Lf\n", p->name, p->n,
+    printf("%s: n=%ld min=%Lf avg=%Lf max=%Lf total=%Lf\n", p->name, (long)p->n,
           (long double)(p->min) / (long double)tsc_freq,
           p->avg / (long double)tsc_freq, (long double)(p->max) / (long double)tsc_freq, p->n * ((long double)(p->avg) / (long double)tsc_freq));
     return 0;
